Timestamp: Make nanoseconds unsigned
Rounding toward zero is not the usual way to split a decimal into integer and fractional part.

Using euclidian division instead causes nanoseconds to always be positive, which means we can store them in u64 instead if i64.
And while we are changing that, we can also switch them to u32 because nanoseconds can't be greater than 10^9 (and u32 supports over 4.10^9)

This also means that seconds of negative timestamps are now decremented by 1. And in particular, the 0 second now only maps to the range [0, 1) instead of (-1, 1).

+++ ar_row/src/lib.rs
@@ -36,7 +36,7 @@ pub mod row_iterator;
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Default)]
 pub struct Timestamp {
     pub seconds: i64,
-    pub nanoseconds: i64,
+    pub nanoseconds: u32,
 }
 
 /// Scale-less decimal number
